more fixes

+++ plugins/pdns/init.rb
@@ -123,8 +123,8 @@ module Optopus
               #event.save!
             end
           elsif ip_record && hostname_record.nil? && !autoupdate_settings['hostname_regex'].nil?
-            hostname_regex = Regexp.new(autoupdate_settings['hostname_regex'])
-            if hostname_regex.match(node.hostname)
+            hostname_regex = autoupdate_settings['hostname_regex']
+            if hostname_regex.match(node.hostname) && hostname_regex.match(ip_record['name'])
               domain = pdns_client.domain_from_name(node.facts['domain'])
               pdns_client.delete_record(ip_record['id'])
               pdns_client.create_record(
@@ -134,13 +134,19 @@ module Optopus
                 :content   => "#{node.facts['ipaddress']}",
                 :ttl       => "600"
               )
+              #log.warn("Record for IP #{node.facts['ipaddress']} already exists for #{ip_record[:name]}; deleting and replacing with #{node.hostname}.")
+              event = Optopus::Event.new
+              event.message = "WARNING: IP #{node.facts['ipaddress']} already exists for #{ip_record["name"]}; deleting and replacing with #{node.hostname}"
+              event.type = 'dns_replace_record'
+              event.properties['node_id'] = node.id
+              event.save!
+            else
+              event = Optopus::Event.new
+              event.message = "WARNING: IP #{node.facts['ipaddress']} already exists for #{ip_record["name"]}, but this record is not a host record. Skipping."
+              event.type = 'dns_replace_record_failed'
+              event.properties['node_id'] = node.id
+              event.save!
             end
-            #log.warn("Record for IP #{node.facts['ipaddress']} already exists for #{ip_record[:name]}; deleting and replacing with #{node.hostname}.")
-            event = Optopus::Event.new
-            event.message = "WARNING: IP #{node.facts['ipaddress']} already exists for #{ip_record["name"]}; deleting and replacing with #{node.hostname}"
-            event.type = 'dns_replace_record'
-            event.properties['node_id'] = node.id
-            event.save!
           elsif hostname_record
             if !hostname_record['content'].eql? node.facts['ipaddress']
               old_ip = hostname_record['content']
